+++ src/incydr/models.py
@@ -41,12 +41,12 @@ from _incydr_sdk.watchlists.models.responses import IncludedDepartmentsList
 from _incydr_sdk.watchlists.models.responses import IncludedDirectoryGroup
 from _incydr_sdk.watchlists.models.responses import IncludedDirectoryGroupsList
 from _incydr_sdk.watchlists.models.responses import IncludedUsersList
-from _incydr_sdk.watchlists.models.responses import Watchlist
 from _incydr_sdk.watchlists.models.responses import WatchlistActor
 from _incydr_sdk.watchlists.models.responses import WatchlistMembersList
 from _incydr_sdk.watchlists.models.responses import WatchlistMembersListV2
-from _incydr_sdk.watchlists.models.responses import WatchlistsPage
+from _incydr_sdk.watchlists.models.responses import WatchlistsPageV2
 from _incydr_sdk.watchlists.models.responses import WatchlistUser
+from _incydr_sdk.watchlists.models.responses import WatchlistV2
 
 
 __all__ = [
@@ -82,8 +82,8 @@ __all__ = [
     "DepartmentsPage",
     "DirectoryGroupsPage",
     "DirectoryGroup",
-    "Watchlist",
-    "WatchlistsPage",
+    "WatchlistV2",
+    "WatchlistsPageV2",
     "WatchlistMembersListV2",
     "WatchlistMembersList",
     "IncludedActorsList",